MSHeatmapSTIP

Generate heatmap for multi-sample STIP result

Usage

MSHeatmapSTIP(x, gl, annotation, interval, ...)

Arguments

x

A gene expression matrix

gl

Marked gene list

annotation

Annotation matrix for genes in expression matrix

interval

A list contains zero points for each gene in each sample

...

parameters passed to Heatmap

Value

A ComplexHeatmap object

Details

Input a gene expression matrix and annotation matrixes, output a heatmap

Author

Zhicheng Ji, Changxin Wan, Beijie Ji